wt88047: Add GPS overlay
xtra server url need to be set, it is empty
in framework, without those xtra server set,
it will not download the gps almanacs,
ephemerides data from wifi or radio network. It will
need to be download from gps satellites directly
which is very slow (> 10 minutes if download from a
single satellite).

Change-Id: Ica79c09288b0e9a0a3c099751db07aee39af974c

Signed-off-by: RazorSharp <kirananto@gmail.com>